+++ datax-example/datax-example-core/src/main/java/com/alibaba/datax/example/util/PathUtil.java
