+++ datax-example/datax-example-core/src/main/resources/example/conf/core.json
