+++ datax-example/datax-example-core/src/test/resources/pathTest.json
@@ -1,0 +1,1 @@
+{}
