+++ datax-example/datax-example-neo4j/src/test/resources/streamreader2neo4j.json
@@ -1,0 +1,51 @@
+{
+  "job": {
+    "content": [
+      {
+        "reader": {
+          "name": "streamreader",
+          "parameter": {
+            "sliceRecordCount": 10,
+            "column": [
+              {
+                "type": "string",
+                "value": "StreamReader"
+              },
+              {
+                "type": "string",
+                "value": "1997"
+              }
+            ]
+          }
+        },
+        "writer": {
+          "name": "neo4jWriter",
+          "parameter": {
+            "uri": "bolt://localhost:7687",
+            "username":"neo4j",
+            "password":"Test@12343",
+            "database":"neo4j",
+            "cypher": "unwind $batch as row CALL apoc.cypher.doIt( 'create (n:`' + row.Label + '`{id:$id})' ,{id: row.id} ) YIELD value RETURN 1 ",
+            "batchDataVariableName": "batch",
+            "batchSize": "3",
+            "properties": [
+              {
+                "name": "Label",
+                "type": "string"
+              },
+              {
+                "name": "id",
+                "type": "STRING"
+              }
+            ]
+          }
+        }
+      }
+    ],
+    "setting": {
+      "speed": {
+        "channel": 5
+      }
+    }
+  }
+}
