+++ datax-example/datax-example-streamreader/src/test/resources/stream2stream.json
