+++ datax-example/pom.xml
@@ -10,6 +10,12 @@
     </parent>
 
     <artifactId>datax-example</artifactId>
+    <packaging>pom</packaging>
+    <modules>
+        <module>datax-example-core</module>
+        <module>datax-example-streamreader</module>
+        <module>datax-example-neo4j</module>
+    </modules>
 
     <properties>
         <maven.compiler.source>8</maven.compiler.source>
@@ -26,17 +32,6 @@
         <dependency>
             <groupId>com.alibaba.datax</groupId>
             <artifactId>datax-core</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
-<!--        默认引入streamwriter,streamreader-->
-        <dependency>
-            <groupId>com.alibaba.datax</groupId>
-            <artifactId>streamwriter</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>com.alibaba.datax</groupId>
-            <artifactId>streamreader</artifactId>
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
         <dependency>
