--- datax-example/src/test/java/com/alibaba/datax/example/streamreader/StreamReader2StreamWriterTest.java
@@ -1,19 +1,0 @@
-package com.alibaba.datax.example.streamreader;
-
-import com.alibaba.datax.example.ExampleContainer;
-import com.alibaba.datax.example.util.PathUtil;
-import org.junit.Test;
-
-/**
- * {@code Author} FuYouJ
- * {@code Date} 2023/8/14 20:16
- */
-
-public class StreamReader2StreamWriterTest {
-    @Test
-    public void testStreamReader2StreamWriter() {
-        String path = "/job/stream/stream2stream.json";
-        String jobPath = PathUtil.getAbsolutePathFromClassPath(path);
-        ExampleContainer.start(jobPath);
-    }
-}
