--- datax-example/src/test/resources/example/conf/core.json
@@ -1,60 +1,0 @@
-{
-    "entry": {
-        "jvm": "-Xms1G -Xmx1G",
-        "environment": {}
-    },
-    "common": {
-        "column": {
-            "datetimeFormat": "yyyy-MM-dd HH:mm:ss",
-            "timeFormat": "HH:mm:ss",
-            "dateFormat": "yyyy-MM-dd",
-            "extraFormats":["yyyyMMdd"],
-            "timeZone": "GMT+8",
-            "encoding": "utf-8"
-        }
-    },
-    "core": {
-        "dataXServer": {
-            "address": "http://localhost:7001/api",
-            "timeout": 10000,
-            "reportDataxLog": false,
-            "reportPerfLog": false
-        },
-        "transport": {
-            "channel": {
-                "class": "com.alibaba.datax.core.transport.channel.memory.MemoryChannel",
-                "speed": {
-                    "byte": -1,
-                    "record": -1
-                },
-                "flowControlInterval": 20,
-                "capacity": 512,
-                "byteCapacity": 67108864
-            },
-            "exchanger": {
-                "class": "com.alibaba.datax.core.plugin.BufferedRecordExchanger",
-                "bufferSize": 32
-            }
-        },
-        "container": {
-            "job": {
-                "reportInterval": 10000
-            },
-            "taskGroup": {
-                "channel": 5
-            },
-            "trace": {
-                "enable": "false"
-            }
-
-        },
-        "statistics": {
-            "collector": {
-                "plugin": {
-                    "taskClass": "com.alibaba.datax.core.statistics.plugin.task.StdoutPluginCollector",
-                    "maxDirtyNumber": 10
-                }
-            }
-        }
-    }
-}
