--- datax-example/src/test/resources/job/stream/stream2stream.json
@@ -1,36 +1,0 @@
-{
-  "job": {
-    "content": [
-      {
-        "reader": {
-          "name": "streamreader",
-          "parameter": {
-            "sliceRecordCount": 10,
-            "column": [
-              {
-                "type": "long",
-                "value": "10"
-              },
-              {
-                "type": "string",
-                "value": "hello，你好，世界-DataX"
-              }
-            ]
-          }
-        },
-        "writer": {
-          "name": "streamwriter",
-          "parameter": {
-            "encoding": "UTF-8",
-            "print": true
-          }
-        }
-      }
-    ],
-    "setting": {
-      "speed": {
-        "channel": 5
-      }
-    }
-  }
-}
